feat: トップ画面を再度、簡易なロゴとテキストに変更し、demo/indexをダッシュボードに変更。
